Finished expenses window, added ability to view expense database, integrated entry of expenses into database
